requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
--extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cpu
llama-cpp-python
pypdf
python-docx
requests
